installed zod package for and added schemas for input forms and parameters; added custome errors and error middleware for them; updated throwing customer errors in services and repo;

+++ src/services/productService.ts
@@ -1,15 +1,31 @@
-import { ProductInputType } from "../types/ProductTypes";
+import { ProductNotFoundError } from "../errors/ProductErrors";
 import { ProductRepository } from "../repositories/productRepository";
+import {
+  CreateProductDTO,
+  UpdateProductDTO,
+} from "../validation/productSchema";
 
-export const createProduct = (product: ProductInputType) => {
+export const createProduct = (product: CreateProductDTO) => {
   return ProductRepository.createProduct(product);
 };
 
 export const deleteProduct = (productId: string) => {
+  const existing = ProductRepository.getProductById(productId);
+
+  if (!existing) {
+    throw new ProductNotFoundError(productId);
+  }
+
   return ProductRepository.deleteProduct(productId);
 };
 
-export const updateProduct = (productId: string, product: ProductInputType) => {
+export const updateProduct = (productId: string, product: UpdateProductDTO) => {
+  const existing = ProductRepository.getProductById(productId);
+
+  if (!existing) {
+    throw new ProductNotFoundError(productId);
+  }
+
   return ProductRepository.updateProduct(productId, product);
 };
